+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - `bd-{color}` Border colors
   - `bd-{nb}, bd-{position}-{nb}, bd-{style}, bd-{position}-{style}` classes
   - `autoClose` option in Dropdown (default: true)
+  - `preventViewport` option in Dropdown (default: false)
   - Axentix.Config with methods : Now, you can register to Axentix your own plugin. (See documentation for more informations)
   - `Axentix.getInstanceByType(type)` utilities | Example: `Axentix.getInstanceByType('Collapsible')`
   - `.spinner-txt` class
@@ -15,7 +16,6 @@
   - Registration of our js components
   - Update definition file
 - BUGFIXES
-  - Dropdown height can exceed viewport size
   - Outline buttons disabled state
 
 # 1.0.0-beta.0 - 2020-07-29